backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
accelerate==0.24.0
scipy==1.11.4
gtts==2.4.0
pydub==0.25.1 